Module Op.Delete

Path deletion.

Path deletion

type t

The type for path deletion operations.

val v_op : id:id -> mark:mark -> created:B00_std.Time.span -> ?⁠post_exec:(op -> unit) -> ?⁠k:(op -> unit) -> B00_std.Fpath.t -> op

v_op declares a path deletion operation, see the corresponding accessors for the semantics of the various arguments.

val v : path:B00_std.Fpath.t -> t

v constructs a bare deletion operation.

val get : op -> t

get o is the delete operation o. Raises Invalid_argument if o is not a delete.

val path : t -> B00_std.Fpath.t

path d is the path to delete.