Module Op.Mkdir

Directory creation.

Directory creation

type t

The type for directory creation operations.

val v_op : id:id -> mark:mark -> mode:int -> created:B00_std.Time.span -> ?⁠post_exec:(op -> unit) -> ?⁠k:(op -> unit) -> B00_std.Fpath.t -> op

v_op declares a directory creation operation, see the corresponding accessors for the semantics of the various arguments.

val v : dir:B00_std.Fpath.t -> mode:int -> t

v constructs a bare mkdir operation.

val get : op -> t

get o is the mkdir o. Raises Invalid_argument if o is not a mkdir.

val dir : t -> B00_std.Fpath.t

dir mk is the directory created by mk.

val mode : t -> int

mode mk are the permissions of the directory created by mk.