Module Op.Notify

End-user notifications.

Notification

type t

The type for notification.

type kind = [

| `End

| `Fail

| `Info

| `Start

| `Warn

]

The type for kinds of notifications.

val v_op : id:id -> mark:mark -> created:B00_std.Time.span -> ?⁠post_exec:(op -> unit) -> ?⁠k:(op -> unit) -> kind -> string -> op

v_op declares a notification operation see the corresponding accessors in Notify for the semantics of the various arguments.

val v : kind:kind -> msg:string -> t

v constructs a notification operation.

val get : op -> t

get o is the notification o. Raise Invalid_argument if o is not a notification.

val kind : t -> kind

kind is the kind of notification.

val msg : t -> string

msg is the message.