Module Op.Wait_files

Waiting on files.

FIXME now that we have Futs maybe we can maybe get rid of this.

type t

The type for wait files operations.

val v_op : id:id -> mark:mark -> created:B00_std.Time.span -> ?post_exec:(op -> unit) -> ?k:(op -> unit) -> B00_std.Fpath.t list -> op

v declares a wait files operation, these are stored in reads.

val v : unit -> t

v constructs a bare wait files operation.