Module B00_cli.Arg

Command line arguments.

Specifying output details

type output_details = [

| `Normal

| `Short

| `Long

]

The type for specifying output detail level.

val output_details : ?docs:string -> ?short_opts:string list -> ?long_opts:string list -> unit -> output_details Cmdliner.Term.t

output_details ~short_opts ~long_opts () are mutually exclusive options to specify short and long output details, without options this is `Normal. short_opts defaults to ["s"; "short"] and long_opts default to ["l"; "long"]. docs is the manual section in which options are documented.