Module B00_fexts

Sets of file extension sets.

File extension sets

type t = B00_std.String.Set.t

The type for sets of file extensions.

type map = B00_std.Fpath.t list B00_std.String.Map.t

The type for files mapped by their extension.

val v : B00_std.Fpath.ext list -> t

v exts is a set for the given extensions.

val ext : B00_std.Fpath.ext -> t

ext e is the set that containes only e.

val find_files : t -> map -> B00_std.Fpath.t list

find_files exts m selects from m the files that have an extension in exts.

val all_files : map -> B00_std.Fpath.t list

all_files m is concatenation of the list of files in m in no particular order.

val exists_file : t -> map -> bool

exists_file exts fm s is true iff find_file fm s is not empty.

val (+) : t -> t -> t

( + ) is String.Set.union.

val (-) : t -> t -> t

( - ) is String.Set.diff.

Constant sets

val all : t

all is the union of all the following sets.

val c_lang : t

c_lang is ".h" and ".c".

val cmark : t

cmark is ".md".

val css : t

css is ".css".

val data : t

data is ".json" and ".xml".

val font : t

font is ".otf", ".ttf", ".woff" and ".woff2".

val html_lang : t

html_lang is ".html", ".css" and ".js"

val image : t

image is ".eps", ".gif", ".ico", ".jpeg", ".jpg", ".pdf", ".png", ".ps" and ".tiff".

val javascript : t

javascript is ".js".

val latex_lang : t

latex_lang is ".tex"; ".sty"; ".bib"; ".bibdoi".

val ocaml_lang : t

ocaml is ".mli", ".ml", ".mld", .mll and .mly.

val sound : t

sound is ".aiff", ".flac", ".mp3" and ".wav".

val tex : t

tex is ".tex".

val video : t

video is ".flv", ".mov" and ".mp4".

val www : t

www is the union of data, font, html_lang, image, sound, video.