Module B00_github.Auth

type t

The type for GitHub authentication.

val v : user:string option -> unit -> (t, string) result

auth ~http ~user () determines authentication via personal access token for user user. It the latter is unspecified it first starts by determining user by looking the

It first looks up the contents of the B0_GITHUB_TOKEN environment variable if that fails it looks up for an existing token in the B00_std.Os.Dir.config () /b0/github/$USER.token file. If that fails instructions are printed on how to setup the token.

val user : t -> string

user a is the GitHub user.

val token : t -> string

token a is the Github token.

Command line interface

val user_env : string

user_env is "B0_GITHUB_USER".

val token_env : string

user_env is "B0_GITHUB_TOKEN".

val envs : Cmdliner.Arg.env list

envs describe user_env and token_env.

val cli : ?opts:string list -> unit -> (t, string) result Cmdliner.Term.t

cli () is a command line interface for GitHub authentication. opts are the options that can be used for specifiying the github user (defaults to ["u";"github"])