Module B00_github.Issue

Repository issues.

Issues

type num = int

The type for issue numbers.

type uri = string

The type for uris.

type t

The type for issues.

val number : t -> num

number i is the issue's number.

val title : t -> string

title i is the issue's title.

val body : t -> string

body t is the issue's body.

val url : t -> string

url i is the issue's url.

val pp : t B00_std.Fmt.t

pp is a formatter for issues.

val pp_short : t B00_std.Fmt.t

pp_short is a short formatter for issues.

val list : B00_http.Httpr.t -> auth -> Repo.t -> (int * t list, string) Stdlib.result

list auth repo lists the issues for repository repo. The integer is the total number of issues.

Requests

val create : B00_http.Httpr.t -> auth -> Repo.t -> title:string -> body:string -> unit -> (num * uri, string) Stdlib.result

create auth repo opens an issue on the repository repo with the given title and body.

val close : B00_http.Httpr.t -> auth -> Repo.t -> num -> (num * uri, string) Stdlib.result

close auth repo n closes issues n on the repository repo