Module B00_github.Release

Repository releases.

Releases

type t

The type for releases.

val id : t -> int

id r is the id of the release.

val tag_name : t -> string

tag_name r is the tag name of the release.

val body : t -> string

body r is the description of the release.

val html_url : t -> string

html_url t is the issue's HTML url.

val assets_url : t -> string

assets_url t is the issue's assets url.

val pp : t B00_std.Fmt.t

pp is a formatter for issues.

val pp_short : t B00_std.Fmt.t

pp_short is a short formatter for issues.

Requests

val create : B00_http.Httpr.t -> auth -> Repo.t -> tag_name:string -> body:string -> unit -> (t, string) Stdlib.result

create auth repo ~tag_name ~body () creates a new release in repository repo with given tag_name and body description.

val get : B00_http.Httpr.t -> auth -> Repo.t -> tag_name:string -> unit -> (t, string) Stdlib.result

get auth repo ~tag_name () gets the release with given tag_name in repo tag_name.

val upload_asset : B00_http.Httpr.t -> auth -> Repo.t -> t -> content_type:string -> name:string -> string -> (unit, string) Stdlib.result

upload_asset auth repo r ~content_type ~name asset uploads assets content asset with file name name and content type content_type in release r of repo r.