Module B00_jsoo

B00 js_of_ocaml support.

This models the js_of_ocaml tool.

Js_of_ocaml

type source_map = [ `Inline | `File ] option

The type for specifying source maps desires.

val tool : B00.Tool.t

tool is the js_of_ocaml executable.

val build_runtime : B00.Memo.t -> opts:B00_std.Cmd.t -> jss:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

build_runtime m ~jss o writes a standalone runtime with JavaScript files jss to o.

val compile : B00.Memo.t -> opts:B00_std.Cmd.t -> source_map:source_map -> jss:B00_std.Fpath.t list -> byte:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

compile m ~source_map ~jss ~byte ~o compiles the JavaScript files jss and byte code object or executable byte to the JavaScript file o.

val link : B00.Memo.t -> opts:B00_std.Cmd.t -> source_map:source_map -> jss:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

link m ~opts ~jss ~o links the JavaScript files jss to o with options opts.

val write_page : ?lang:string -> ?generator:string -> ?styles:string list -> ?scripts:string list -> ?title:string -> B00.Memo.t -> o:B00_std.Fpath.t -> unit

write_page m ~title ~o writes to file o a full HTML document whose body contains only a B00_htmlg.El.noscript element that entices the user, in english, to enable JavaScript. title defaults to the basename of o without its extension, for the other arguments and more information see B00_htmlg.El.basic_page.