Module B00_ocaml.Archive

Archiving.

Tool invocations for archiving.

val cstubs : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> c_objs:B00_std.Fpath.t list -> odir:B00_std.Fpath.t -> oname:string -> unit

cstubs m ~conf ~opts ~c_objs ~odir ~oname creates in directory odir C stubs archives for a library named oname.

val byte : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> has_cstubs:bool -> cobjs:B00_std.Fpath.t list -> odir:B00_std.Fpath.t -> oname:string -> unit

byte_archive m ~opts ~has_cstubs ~cobjs ~obase creates in directory odir a bytecode archive named oname with the OCaml bytecode compilation objects cobjs.

val native : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> has_cstubs:bool -> cobjs:B00_std.Fpath.t list -> odir:B00_std.Fpath.t -> oname:string -> unit

native m ~opts ~has_cstubs ~cobjs ~obase creates in directory odir a native code archive named oname with the OCaml native code compilation objects cobjs.

val code : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> code:Conf.code -> has_cstubs:bool -> cobjs:B00_std.Fpath.t list -> odir:B00_std.Fpath.t -> oname:string -> unit

archive is byte_archive or native_archive according to code.

val native_dynlink : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> has_cstubs:bool -> cmxa:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit