Module B00_ocaml.Link

OCaml linking.

module Deps : sig ... end

Link dependencies.

val byte_exe : ?⁠post_exec:(B000.Op.t -> unit) -> ?⁠k:(int -> unit) -> ?⁠args:B00_std.Cmd.t -> B00.Memo.t -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

byte_exe m ~args ~c_objs ~cmos ~o links the C objects c_objs and the OCaml compilation object files cobjs into a byte code executable o compiled in -custom mode.

val native_exe : ?⁠post_exec:(B000.Op.t -> unit) -> ?⁠k:(int -> unit) -> ?⁠args:B00_std.Cmd.t -> B00.Memo.t -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

byte_exe m ~args ~c_objs ~cobjs ~o links the C objects c_objs and the OCaml compilation object files cobjs into a native code executable o. An include is added to each element of cobjs in order to lookup potential C stubs.

val exe : ?⁠post_exec:(B000.Op.t -> unit) -> ?⁠k:(int -> unit) -> ?⁠args:B00_std.Cmd.t -> B00.Memo.t -> code:Cobj.code -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

exe is byte_exe or native_exe according to code.