Module Mod.Ref

Module digested references.

TODO. Use that in B00_odoc.

Module references

type t

The type for module references as found in compilation objects. This is a module name and a digest of its interface.

val v : string -> Digest.t -> t

v n d is a module reference with name n and digest d.

val name : t -> Name.t

name m is the capitalized module name of module reference m.

val digest : t -> Digest.t

digest m is the interface digest of module reference m.

val equal : t -> t -> bool

equal m m' is true iff m and m' are the same reference.

val compare : t -> t -> int

compare m m' is a total order on module references.

val pp : t B00_std.Fmt.t

pp formats a module reference.

module Set : sig ... end

Module reference sets.

module Map : sig ... end

Module reference maps.