Module B00_ocaml.Ocamlpath

OCAMLPATH search path.

val get : B00.Memo.t -> B00_std.Fpath.t list option -> B00_std.Fpath.t list B00.Memo.fiber

get m o k is k ps if o is Some ps and otherwise in order:

If the OCAMLPATH environment variable is defined in m and non-empty its content is parsed s according to B00_std.Fpath.list_of_search_path.

If the OPAM_SWITCH_PREFIX environment variable is defined with a path P then [P/lib] is used.

The fiber fails.