Module B00_odoc.Compile

Compiling .cmi, .cmti, .cmt and .mld files to .odoc files.

Compile

module Dep : sig ... end

Compilation dependencies, for compiling .cmi, .cmti or .cmt files to .odoc files.

module Writes : sig ... end

Compilation writes, files written by compiling a .cmi, .cmti, .cmt or .mld file to an .odoc file.

val cmd : ?⁠resolve_forward_deps:bool -> ?⁠hidden:bool -> B00.Memo.t -> odoc_deps:B00_std.Fpath.t list -> writes:B00_std.Fpath.t list -> pkg:string -> B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

cmd m ~resolve_forward_deps ~hidden ~odoc_deps ~writes ~pkg cobj o compiles the compilation object cobj part of package pkg to destination odoc file o.

odoc_deps are the odoc files the compilation of cobj reads. They can be obtained by resolving the result of Dep on cobj.

writes specifies the writes of the command they can be obtained via Writes.

resolve_forward_deps and hidden are the corresponding odoc options. See odoc compile --help.

Convenience

val to_odoc : B00.Memo.t -> ?⁠hidden:bool -> pkg:string -> odoc_deps:B00_std.Fpath.t list -> B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

to_odoc m ~hidden ~pkg ~odoc_deps obj ~o compiles obj (which can be any of a .cmi, .cmt, .cmti or .mld file) to an odoc file o assuming it depends on odoc_deps and is part of package pkg.