Module Compile.Writes

Compilation writes, files written by compiling a .cmi, .cmti, .cmt or .mld file to an .odoc file.

As determined by the odoc compile-targets command.

Compilation file writes

val write : B00.Memo.t -> B00_std.Fpath.t -> to_odoc:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m cobj ~to_odoc ~o writes the files written by a compilation of cobj to to_odoc to o.

val read : B00.Memo.t -> B00_std.Fpath.t -> B00_std.Fpath.t list B00.Memo.fiber

read m file reads the result of a write from file and continues with the files that will be written.