Module Html.Dep

HTML generation dependencies, for compiling a package (sic) to HTML files.

As returned by the odoc html-deps command. These dependencies need to be resolved to concrete .odoc files by some external mean.

Dependencies

type t

The type for HTML generation dependencies.

val pkg : t -> string

pkg d is the package of d.

val name : t -> string

name d is the module name of d.

val digest : t -> Digest.t

digest d is the digest of d.

val to_compile_dep : t -> Compile.Dep.t

to_compile_dep d is d as a compilation depencency (simply drops the pkg).

val write : B00.Memo.t -> odoc_files:B00_std.Fpath.t list -> B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m ~odoc_files pkg_odoc_dir o writes the odoc dependencies of the package directory pkg_odoc_dir that contains the odoc files odoc_files to o.

val read : B00.Memo.t -> B00_std.Fpath.t -> t list B00_std.Fut.t

read m file reads the result of a Dep.write from file and continues with the dependencies.