Module Html.Writes

HTML generation writes, files written by generating HTML files from an .odoc file.

As determined by the odoc html-targets command.

HTML generation file writes

val write : B00.Memo.t -> odoc_deps:B00_std.Fpath.t list -> B00_std.Fpath.t -> to_dir:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m ~odoc_deps odoc ~to_dir ~o writes to o the files written by an HTML generation of the .odoc file odoc to to_dir.

odoc_deps are the .odoc file dependencies for the odoc file, they can be obtained by resolving the result of Dep on the package odoc directory of odoc.

val read : B00.Memo.t -> B00_std.Fpath.t -> B00_std.Fpath.t list B00.Memo.fiber

read m file reads the result of a write from file and continues with the files that will be written.