Module Support_files.Writes

Support files writes.

As determined by the odoc support-files-targets command.

Support files writes

val write : ?without_theme:bool -> B00.Memo.t -> to_dir:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m ~without_theme ~to_dir ~o writes to o the support files written to the to_dir directory. without_theme is the corresponding odoc option.

val read : B00.Memo.t -> B00_std.Fpath.t -> B00_std.Fpath.t list B00_std.Fut.t

read m file reads the result of a Writes.write from file and continues with the files that will be written.