Module B00_odoc.Theme

Odoc theme support.

Themes names

type name = string

The type for theme names.

val odoc_default : name

odoc_default is the default odoc theme ("odoc.default").

val odig_default : name

odig_default is the default odig theme ("odig.default").

val default_uri : string

default_uri is a default uri to give to Html.cmd, this is "_odoc-theme". This will use that directory in the html directory.

User preference

val config_file : B00_std.Fpath.t

config_file is the file relative to the user's Os.Dir.config directory for specifying the odoc theme.

val get_user_preference : unit -> (name option, string) Stdlib.result

get_user_preference () is the user prefered theme name (if any).

val set_user_preference : name option -> (unit, string) Stdlib.result

set_user_preference t sets the user prefered theme to t.

Themes

type t

The type for themes.

val name : t -> name

name t is the theme name.

val path : t -> B00_std.Fpath.t

path t is the path to the theme directory.

val pp_name : t B00_std.Fmt.t

pp_name formats a theme's name.

val pp : t B00_std.Fmt.t

pp formats a theme.

Queries

val of_dir : B00_std.Fpath.t -> t list

of_dir sharedir are the themes found in sharedir. These are formed by looking up in sharedir for directory paths of the form PKG/odoc-theme/ID/ in sharedir which yields a theme named by PKG.ID.

val find : fallback:name option -> name -> t list -> (t, string) Stdlib.result

find ~fallback n ts finds theme n in ts. An error is returned if n cannot be found, in that case it indicates fallback will be used instead (if specified).

Writing

val write : B00.Memo.t -> t -> to_dir:B00_std.Fpath.t -> unit

write m t ~to_dir writes theme to directory to_dir. This simply copies path to to_dir.