Module B00_trace.Compilation_database

JSON compilation database.

Output build operation spawns in JSON compilation database format (see also here).

The format is a bit underpowered (see also the discusion in the merlin issue tracker here) and doesn't exactly match B0's build model here are a few notes on the treatment:

The file field. B0's operations do not identify a "main source" processed by a step. For now we use the first element of Op.reads TODO maybe we should spell out this convention in Memo.

The output field. B0's operations support multiple writes. If multiple files are written we repeat the command for each write. We add an numbered id field to indicate that this is the same command.

No specification of the environment, we add it under an env key as an array of strings.

We never generate command, we only use arguments.

JSON compilation database

val of_ops : B000.Op.t list -> B00_serialk_json.Jsong.t