Module B00_ui.Cli

Cmdliner fragments.

Specifying output detail

type out_details = [

| `Normal

| `Short

| `Long

]

The type for specifying output detail level.

val out_details : ?⁠docs:string -> ?⁠short_opts:string list -> ?⁠long_opts:string list -> unit -> out_details Cmdliner.Term.t

out_details ~short_opts ~long_opts () are mutually exclusive options to specify short and long output details, without options this is `Normal. short_opts defaults to ["s"; "short"] and long_opts default to ["l"; "long"]. docs is the manual section in which options are documented.