Module B0_b0.Def

B0_def generic support.

Generic implementation of a few standard commands we need for B0 defintions. FIXME this could likely be moved to a B0_driver_ui in B0 care.

val list : (module B0_def.S) -> B0_driver.Conf.t -> B00_ui.Cli.out_details -> string list -> B0_driver.Exit.t

list (module Def) c details ns lists definition Def named ns with details details. If ns is empty all definitions are listed.

val edit : (module B0_def.S) -> B0_driver.Conf.t -> string list -> B0_driver.Exit.t

edit (module Def) c ns edits the B0 files which define Defs named ns. If ns is empty all the B0 files that have definitions of kind Def are edited.

val get : (module B0_def.S) -> B0_driver.Conf.t -> B00_ui.Cli.out_details -> string -> string list -> B0_driver.Exit.t

get (module Def) k ns gets key k in the metadata of definitions named ns with details details. If ns is empty all definitions are listed.