Module type B0_def.VALUE

The type for values to be named.

This just indicates an identifier for the kind of names, how to retreive the VALUE.def value in a value and how to format the names.

Values

type t

The type of defined values.

val def_kind : string

def_kind is the kind of defined value, uncapitalized.

val def : t -> def

def v is v's definition.

val pp_name_str : string B00_std.Fmt.t

pp_name_str pretty prints value names.