Module B0_driver.Env

Driver environment variables.

val b0_dir : string

b0_dir is the variable used to specify the b0 directory.

val b0_file : string

b0_file is the variable used to specify the b0 file to use.

val cache_dir : string

cache_dir is the variable used to specify the b0 cache directory.

val color : string

color is the variable used to specify tty styling.

val code : string

code is the variable used to specify the code to which the driver compiles to.

val hash_fun : string

hash_fun is the variable used to specify the cache hash function.

val jobs : string

jobs is the variable used to specify the maximal number of spawns.

val verbosity : string

verbosity is the variable used to specify log verbosity.