Module B0_release.Meta

Metadata keys for releases.

val src_archive_name : string B0_meta.key

archive_name is the basename of the source release's archive.

val src_archive_ext : B00_std.Fpath.ext B0_meta.key

archive_ext is the file extension of the source release's archive.

val src_archive_url : string B0_meta.key

src_archive_url is an URL pattern that specifies a source release on the WWW. The following variables are to be subsituted %‌%ARCHIVE_NAME%‌%, %‌%ARCHIVE_EXT%‌%, %‌%VERSION%‌%, %‌%VERSION_NUM%‌%, see src_archive_url_of_pack for more details.