Module B0_unit.Action

Program execution

val exec_cwd : (build -> t -> B0_std.Fpath.t B0_std.Fut.t) B0_meta.key

exec_cwd is a function to determine a current working directory for exec actions.

val scope_cwd : build -> t -> B0_std.Fpath.t B0_std.Fut.t

scope_cwd is function that determines the unit's scope.

val exec_env : (build -> t -> B0_std.Os.Env.assignments B0_std.Fut.t) B0_meta.key

exec_env is a function to determine an environement for exec actions.

val exec : action

exec is an action that B0_std.Os.Exit.exec's a unit's outcome as follows:

The executed file is the unit's B0_meta.exe_file.

The arguments have B0_std.Fpath.basename of B0_meta.exe_file as the program name and the action's args as arguments.

If the unit defines exec_cwd, it is used to determine the cwd otherwise the default of B0_std.Os.Exit.exec is used.

If the unit defines exec_env, it is used to determine the environment otherwise the default B0_std.Os.Exit.exec is used.

val exec_file : build -> t -> B0_std.Fpath.t -> B0_std.Cmd.t -> B0_std.Os.Exit.t B0_std.Fut.t

exec_file u file argv is a B0_std.Os.Exit.exec with file and argv and:

If the unit u defines exec_cwd, it is used to determine the cwd otherwise the default of B0_std.Os.Exit.exec is used.

If the unit u defines exec_env, it is used to determine the environment otherwise the default B0_std.Os.Exit.exec is used.