Module B00.Env

Build environment.

Build environments specify the environment of tool spawns.

TODO Now that tool lookup moved to Memo, is it still worth sense to have that separate ?

Environment

type t

The type for build environments.

val v : ?forced_env:B00_std.Os.Env.t -> B00_std.Os.Env.t -> t

v ~lookup ~forced_env env is a build environment with:

forced_env is environment forced on any tool despite what it declared to access, defaults to Os.Env.empty

env the environment read by the tools' declared environment variables.

val env : t -> B00_std.Os.Env.t

env e is e's available spawn environment.

val forced_env : t -> B00_std.Os.Env.t

forced_env e is e's forced spawn environment.