Module Op.Write

File writes.

File writes

type t

The type for file write operations.

val v_op : id:id -> mark:mark -> created:B00_std.Time.span -> ?post_exec:(op -> unit) -> ?k:(op -> unit) -> stamp:string -> reads:B00_std.Fpath.t list -> mode:int -> write:B00_std.Fpath.t -> (unit -> (string, string) result) -> op

write declares a file write operations, see the corresponding accessors for the semantics of the various arguments.

val v : stamp:string -> mode:int -> file:B00_std.Fpath.t -> data:(unit -> (string, string) result) -> t

v constructs a bare write operation.

val get : op -> t

geo o is the write o. Raises Invalid_argument if o is not a write.

val stamp : t -> string

stamp w is the file write stamp used for caching.

val mode : t -> int

int w is the mode of the file written by w.

val file : t -> B00_std.Fpath.t

file w is the file written by w.

val data : t -> (string, string) result

data w invokes and discards the write data function. If the write data function raises this is turned into an Error _.

val discard_data : t -> unit

discard_data w discards the write data function.