Module B00_base64

Base64 codec.

Codecs Base64 according to RFC 4684.

val encode : string -> string

encode s is the Base64 encoding of s.

val decode : string -> (string, int) result

decode s is the Base64 decode of s. In case of error the integer indicates the byte index of the error for an invalid alphabet character error or the length of the string if the string length is not a multiple of 4.