Module B00_github.Repo

GitHub repositories.

type t

The type for GitHub repositories.

val v : owner:string -> string -> t

repo ~owner name identifiers a GitHub repository.

val of_url : string -> (t, string) result

of_url url parses an owner and repo name from the first two segments of url's path.

val owner : t -> string

owner r is r's owner.

val name : t -> string

name r is r's name.

val req_json_v3 : ?headers:B00_http.Http.headers -> B00_http.Httpr.t -> Auth.t -> t -> path:string -> B00_http.Http.meth -> v3_body -> (B00_serialk_json.Json.t, string) result

req_json_v3 is like B00_github.req_json_v3 but performs given the root subpath on the given repo.

val query_v4 : B00_http.Httpr.t -> Auth.t -> t -> string -> (B00_serialk_json.Json.t, string) result

query_v4 auth r q performs the subgraph query q on repo r using authentication auth.