Module B00_ocaml.Lib

Libraries

type t

The type for libraries.

val name : t -> Lib_name.t

name l is the library name of l.

val cmis : t -> B00_std.Fpath.t list

cmis l is the list of cmis for the library.

val dir : t -> B00_std.Fpath.t

dir l is the path to the library directory. Incidentally this is also the path to the library includes.

val archive : code:Cobj.code -> t -> B00_std.Fpath.t

archive ~code l is the path to the library archive for code c. Not checked for existence.