Module B00_ocaml.Lib_resolver

Library resolvers.

type t

The type for library resolvers.

val create : B00.Memo.t -> memo_dir:B00_std.Fpath.t -> ocamlpath:B00_std.Fpath.t list -> t

create m ocamlpath is a library resolver looking for libraries in the OCAMLPATH ocamlpath.

val find : t -> Lib_name.t -> Lib.t B00.Memo.fiber

find r l finds library l using r. The fiber fails if the library name cannot be found.

This should look a list and sort.