Module B00_ocaml.Link

Linking.

Tool invocations for linking.

val byte : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

byte_exe m ~opts ~c_objs ~cmos ~o links the C objects c_objs and the OCaml compilation object files cobjs into a byte code executable o compiled in -custom mode.

val native : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

byte_exe m ~opts ~c_objs ~cobjs ~o links the C objects c_objs and the OCaml compilation object files cobjs into a native code executable o. An include is added to each element of cobjs in order to lookup potential C stubs.

val code : ?post_exec:(B000.Op.t -> unit) -> ?k:(int -> unit) -> B00.Memo.t -> conf:Conf.t -> opts:B00_std.Cmd.t -> code:Conf.code -> c_objs:B00_std.Fpath.t list -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

code is byte or native according to code.