Module Link.Deps

Link dependencies.

val write : B00.Memo.t -> cobjs:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

write m cobjs o writes the external link dependencies and order of compilation unit objects of cobjs to o.

val read : B00.Memo.t -> B00_std.Fpath.t -> (Cobj.t list * Mod_ref.Set.t) B00.Memo.fiber

read m file has the cobjs of a write to file in dependency order and external dependencies to resolve.