Module B00_odoc

B0 odoc support.

Note. This models the odoc tool and its invocations. Dependecy resolution is not part of the module yet we could try to abstract it via a few functions and provide a full driver story.

Odoc

val tool : B00.Tool.t

tool is the odoc tool.

module Compile : sig ... end

Compiling .cmi, .cmti, .cmt and .mld files to .odoc files.

module Html : sig ... end

Generate HTML from .odoc files.

module Html_fragment : sig ... end

Generate HTML fragments from .mld files.

module Support_files : sig ... end

Generate support files for the HTML.

module Theme : sig ... end

Odoc theme support.