Module Compile.Dep

Compilation dependencies, for compiling .cmi, .cmti or .cmt files to .odoc files.

As determined by the odoc compile-deps command. These dependencies need to be resolved to concrete .odoc files by some external mean.

Dependencies

type t

The type for odoc compilation dependencies.

val name : t -> string

name d is the module name of d.

val digest : t -> Digest.t

digest d is the digest of d.

val pp : t B00_std.Fmt.t

pp formats a dependency.

val write : B00.Memo.t -> B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m cobj o writes the odoc dependencies of the compilation object cobj to o.

val read : B00.Memo.t -> B00_std.Fpath.t -> t list B00_std.Fut.t

read m file reads the result of a write from file and continues with the dependencies.