Module B00_odoc.Html

Generate HTML from .odoc files.

HTML generation

module Dep : sig ... end

HTML generation dependencies, for compiling a package (sic) to HTML files.

module Writes : sig ... end

HTML generation writes, files written by generating HTML files from an .odoc file.

val cmd : ?hidden:bool -> ?theme_uri:string -> B00.Memo.t -> odoc_deps:B00_std.Fpath.t list -> writes:B00_std.Fpath.t list -> B00_std.Fpath.t -> to_dir:B00_std.Fpath.t -> unit

cmd m ~hidden ~theme_uri ~odoc_deps ~writes odoc ~to_dir generates HTML for the .odoc file odoc to output directory to_dir with the odoc html command.

odoc_deps are the .odoc file dependencies for the odoc file, they can be obtained by resolving the result of Dep on the package odoc directory of odoc.

writes specifies the writes of the command, they can be obtained via Writes.

theme_uri and hidden are the corresponding odoc options. See odoc html --help.

Convenience

val write : B00.Memo.t -> ?theme_uri:string -> html_dir:B00_std.Fpath.t -> odoc_deps:B00_std.Fpath.t list -> B00_std.Fpath.t -> unit

write m ~theme_uri ~html_dir ~odoc_deps odoc writes the html for odoc in html_dir assuming it depends on the odoc_files in odoc_deps.