Module B00_pdf_viewer

PDF viewer interaction.

Environment variables

module Env : sig ... end

Environment variables.

Show PDFs

type t

The type for specifying a PDF viewer.

val find : ?search:B00_std.Fpath.t list -> pdf_viewer:B00_std.Cmd.t option -> unit -> (t option, string) result

find ~search ~pdf_viewer tries to find a PDF viewer in a platform dependent way.

val show : t option -> B00_std.Fpath.t -> (unit, string) result

show pdf_viewer file shows PDF file using the viewer pdf_viewer (if None an error message is returned mentioning no viewer was found).

Cli interaction

val pdf_viewer : ?docs:string -> ?opts:string list -> unit -> B00_std.Cmd.t option Cmdliner.Term.t

pdf_viewer ~docs ~opts () is an option and PDFVIEWER environment variable to use with pdf_viewer argument of find. opts are the cli options and default to "pdf-viewer". docs is the documentation section where the option gets documented.