Module B0_build

B0 builds.

Centralizes the information to run and orchestrate a B0 build.

Build

type t = B0_unit.build

The type for builds.

val memo : t -> B00.Memo.t

memo b the memoizer for the build.

val store : t -> B00.Store.t

store b is the store for the build.

val locked : t -> bool

locked b is true iff b is a locked build. In a locked build build units that are built are fixed before the build starts.

module Unit : sig ... end

Built units.

Creating and running

val create : root_dir:B00_std.Fpath.t -> b0_dir:B00_std.Fpath.t -> B00.Memo.t -> locked:bool -> B0_unit.t list -> t

val run : t -> (unit, unit) Stdlib.result

run b runs the build.