Module B0_driver.Compile

Driver compilation.

Driver compilation is automatically handled by with_b0_file. But a few driving bits are exposed here.

val build_dir : Conf.t -> driver:t -> B00_std.Fpath.t

build_dir c ~driver is a build directory for driver driver in configuration c.

val build_log : Conf.t -> driver:t -> B00_std.Fpath.t

build_log c ~driver is a build log file for driver driver in configuration c.

val exe : Conf.t -> driver:t -> B00_std.Fpath.t

exe c ~driver is the driver executable for driver driver in configuration c.

val compile : Conf.t -> driver:t -> B0_file.t -> (B00_std.Fpath.t, string) result

compile c ~driver b0_file compiles b0_file with driver driver in configuration c. If all is well the executable file path is returned.