Module B0_ocaml.Unit

OCaml build units

Units

val exe : ?⁠doc:string -> ?⁠meta:B0_meta.t -> ?⁠requires:B00_ocaml.Lib_name.t list -> ?⁠name:string -> string -> srcs:B0_srcs.t -> B0_unit.t

exe n is a build unit for an executable named n.

doc is the unit doc string.

meta is the initial metadata.

requires are the OCaml libraries required to compile the executable.

name is the name of the unit (defaults to n).

srcs are the executable sources. All files with extension .ml, .mli, .c and .h are considered for compiling and linking the executable.

val lib : ?⁠doc:string -> ?⁠meta:B0_meta.t -> ?⁠requires:B00_ocaml.Lib_name.t list -> ?⁠name:string -> B00_ocaml.Lib_name.t -> srcs:B0_srcs.t -> B0_unit.t

lib n is a built unit for a library named l.

doc is the unit doc string.

meta is the initial metadata.

requires are the OCaml libraries required to compile the library.

name is the name of the build unit (default to n with . substituted by -)

srcs are the library sources. extension .ml, .mli, .c and .h are considered for compiling and linking the executable.